refactor(config): move runtime_cast to lab harvest (its sole consumer)
runtime_cast had a single consumer (param_decomp_lab/harvest/reservoir.py); move it
there and drop it from param_decomp/base_config.py, which now exposes only BaseConfig +
Probability. Docstrings/README/CLAUDE updated to match.
